ajustes

+++ Lista1/Exercicio05/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         Scanner in = new Scanner(System.in);
         System.out.println("Digite a velocidade em Km/h:");
 
-        vkm = in.nextInt();
+        vkm = in.nextDouble();
 
         vms = vkm / 3.6;
