+++ Lista1/Exercicio08/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         Scanner in = new Scanner(System.in);
         System.out.println("Digite a medida em polegadas:");
 
-        medida = in.nextInt();
+        medida = in.nextDouble();
 
         result = medida * 25.4;
